perf: fix double setImage + OpenGL + Windows timer + playback diagnostics
- Remove redundant 1st setImage() in show_frame_fast when W/L is enabled
  (was causing 2x GPU texture upload per frame)
- Add pg.setConfigOptions(useOpenGL=True) for hardware-accelerated rendering
- Add timeBeginPeriod(1) on Windows to reduce QTimer floor from ~15ms to ~1ms
- Add playback_diagnostics.py module for runtime FPS/memory profiling
- Add test_dicom_multiframe_bench.py: 44 benchmarks for multiframe DICOM pipeline
- Fix test_to_grayscale_array to expect float32 (aligns with SPEC-001)

+++ tests/unit/test_pixel_utils_display_range.py
@@ -69,7 +69,7 @@ def test_to_grayscale_array_preserves_uint16_range() -> None:
     frame = np.array([[0, 4095]], dtype=np.uint16)
     gray = to_grayscale_array(frame)
 
-    assert gray.dtype == np.float64
+    assert gray.dtype == np.float32
     assert gray.max() == pytest.approx(4095.0)
